// maeri/src/app/models/patient.model.ts
export interface Patient {
  phone: string;
  name: string;
  password: string;
  confirm: string;
  ville: string;
  email: string;
  adminEmail: string;
  adminId: string;
  isRetailer: boolean;
  poste: string;
  productsToSale: [];
  quartier: string;
  role: any[];
  storeId: string;
  _id: string;
  created: string;
  age: number;
  sexe: string;
  religion: string;
  nationalite: string;
}