// maeri/src/app/models/user.ts
export class User {
  telephone: string;
  name: string;
  password: string;
  confirm: string;
  ville: string;
  email: string;
  adminEmail: string;
  adminId: string;
  isRetailer: boolean;
  phone: string;
  poste: string;
  productsToSale: [];
  quartier: string;
  role: any[];
  storeId: string;
  _id: string;
}